Rename REST pxl into pixel and rect into rectanlge + move to package be.cytomine.multidim + remove stuff about the conversion (now in core)
