+++ src/groovy/be/cytomine/multidim/hdf5/output/FileReaderCache.groovy
@@ -1,4 +1,4 @@
-package be.charybde.multidim.hdf5.output
+package be.cytomine.multidim.hdf5.output
 
 import java.util.concurrent.Executors
