+++ src/groovy/be/cytomine/multidim/hdf5/output/HDF5FileReader.groovy
@@ -1,6 +1,6 @@
-package be.charybde.multidim.hdf5.output
+package be.cytomine.multidim.hdf5.output
 
-import be.charybde.multidim.exceptions.CacheTooSmallException
+import be.cytomine.multidim.exceptions.CacheTooSmallException
 import ch.systemsx.cisd.hdf5.HDF5Factory
 import ch.systemsx.cisd.hdf5.IHDF5Reader
