+++ src/groovy/be/cytomine/multidim/hdf5/output/HDF5Geometry.groovy
@@ -1,4 +1,4 @@
-package be.charybde.multidim.hdf5.output
+package be.cytomine.multidim.hdf5.output
 
 /**
  * Created by laurent on 08.01.17.
